Source per VMMaker-dtl.339 (VMMaker 4.13.1)
Add primitiveAllObjects, answering an array of all objects in the object
memory at the time of calling the primitive, possibly excluding those that
were garbage collected as a side effect of allocating the result array for
the primitive.

Fix comment for ObjectMemory>>lengthOf:baseHeader:format: to indicate that
length includes the fixed fields as well as indexable fields (the length
of an OrderedCollection instance is 3, not 0).

Add shorten:toIndexableSize: with error checks and correct size logic for
long arrays. Change the return value to be the number of bytes released
to free memory, allowing sender to test for success. The implementation
for ClassicObjectMemory is tested for both 32-bit and 64-bit object memories.
The implementation for NewObjectMemory is based on this, but has not been
tested.

Add primitiveTestShortenIndexableSize to allow shorten:toIndexableSize:
to be exercised from the image. This primitive is for testing purposes
only and will probably be removed in the future.

FT2Plugin to external plugins in VMMakerTool class>>defaultUnixSpec because
some platforms (e.g. OS X) may not have the runtime library conveniently
available.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@2854 fa1542d4-bde8-0310-ad64-8ed1123d492a