VMMaker 4.14.1
Eliot's fix for sqComplexEvent on 64-bit object memory to support iOS VM.

Requires platforms source version SVN 3412 or greater.

Background:

In sq.h, the sqInputEvent struct was originally defined with all fields of type int. These fields are accessed in primitiveGetNextEvent by treating the sqInputEvent struct as an array of int. The sqComplexEvent struct variation was introduced for the iOS VM, and uses one of the available fields to store an object pointer of type usqInt.

In the case of a 64 bit object memory, sizeof(usqInt) is 8, and the object pointer cannot be stored in a C variable of type int. Therefore, change the declaration of all sqInputEvent data structures to have fields of type long, which are large enough to store objects pointers on all platforms of interest. This allows the event struct to safely store both integers and object pointers regardless of object memory word size.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@3413 fa1542d4-bde8-0310-ad64-8ed1123d492a