Do not use -O3 optimization, -O2 is safer and works well.
gcc 4.9.2 gives crashes and heisenbugs with OSPP compiled with -O3.

Symptoms:
(1 to: 10) collect: [ :e | (CommandShell new pipeline: 'ps > /dev/null | cat') output]
  ==> segfaults in some of the spawned child processes

This appears to be an actual compiler bug, although it goes away when print
statements are added, so I cannot say for sure (and it is difficult to attach
gdb to the newly spawned child process before it crashes). The bad behavior
happens only with -O3 and there are no real performance benefits compared
to -O2 (bytecodes faster, sends slower).



git-svn-id: http://squeakvm.org/svn/squeak/trunk@3753 fa1542d4-bde8-0310-ad64-8ed1123d492a